+++ ACMICPC/ACMICPC/01000-02000/01958.c
@@ -25,10 +25,10 @@ int Problem01958(void)
     nLenC = strlen(szC);
 
     ppp_nLCS_DP = (int***)malloc(sizeof(int **)*(nLenA + 1));
-    for(int i = 0; i <= nLenA; i++)
+    for(size_t i = 0; i <= nLenA; i++)
     {
         ppp_nLCS_DP[i] = (int**)malloc(sizeof(int*)*(nLenB + 1));
-        for(int j = 0; j <= nLenB; j++)
+        for(size_t j = 0; j <= nLenB; j++)
         {
             ppp_nLCS_DP[i][j] = (int*)malloc(sizeof(int)*(nLenC + 1));
             memset(ppp_nLCS_DP[i][j], 0, sizeof(int)*(nLenC + 1));
@@ -36,11 +36,11 @@ int Problem01958(void)
     }
 
 
-    for(int i = 1; i < nLenA; i++)
+    for(size_t i = 1; i < nLenA; i++)
     {
-        for(int j = 1; j < nLenB; j++)
+        for(size_t j = 1; j < nLenB; j++)
         {
-            for(int k = 1; k < nLenC; k++)
+            for(size_t k = 1; k < nLenC; k++)
             {
                 if(szA[i] == szB[j] && szB[j] == szC[k])
                 {
@@ -64,9 +64,9 @@ int Problem01958(void)
     nLenResult = ppp_nLCS_DP[nLenA - 1][nLenB - 1][nLenC - 1];
     printf("%zd\n", nLenResult);
 
-    for(int i = 0; i <= nLenA; i++)
+    for(size_t i = 0; i <= nLenA; i++)
     {
-        for(int j = 0; j <= nLenB; j++)
+        for(size_t j = 0; j <= nLenB; j++)
         {
             free(ppp_nLCS_DP[i][j]);
         }
