+++ ACMICPC/ACMICPC/02000-03000/02216.c
@@ -24,24 +24,24 @@ int Problem02216(void)
     nLenB = strlen(szB);
 
     pp_nScore_DP = (int **)malloc(sizeof(int*)*nLenA);
-    for(int i = 0; i < nLenA; i++)
+    for(size_t i = 0; i < nLenA; i++)
     {
         pp_nScore_DP[i] = (int*)malloc(sizeof(int)*nLenB);
         memset(pp_nScore_DP[i], 0, sizeof(int)*nLenB);
     }
 
-    for(int i = 0; i < nLenA; i++)
+    for(size_t i = 0; i < nLenA; i++)
     {
         pp_nScore_DP[i][0] = i * nScoreBlank;
     }
-    for(int i = 0; i < nLenB; i++)
+    for(size_t i = 0; i < nLenB; i++)
     {
         pp_nScore_DP[0][i] = i * nScoreBlank;
     }
 
-    for(int i = 1; i < nLenA; i++)
+    for(size_t i = 1; i < nLenA; i++)
     {
-        for(int j = 1; j < nLenB; j++)
+        for(size_t j = 1; j < nLenB; j++)
         {
             if(szA[i] == szB[j])
             {
@@ -56,7 +56,7 @@ int Problem02216(void)
     }
     printf("%d\n", pp_nScore_DP[nLenA-1][nLenB-1]);
     
-    for(int i = 0; i < nLenA; i++)
+    for(size_t i = 0; i < nLenA; i++)
     {
         free(pp_nScore_DP[i]);
     }
