+++ ACMICPC/ACMICPC/05000-06000/05582.c
@@ -17,16 +17,16 @@ int Problem05582(void)
     nLenB = strlen(szB);
 
     pp_nAB_DP = (int**)malloc(sizeof(int *)*nLenA);
-    for(int i = 0; i < nLenA; i++)
+    for(size_t i = 0; i < nLenA; i++)
     {
         pp_nAB_DP[i] = (int*)malloc(sizeof(int)*nLenB);
         memset(pp_nAB_DP[i], 0, sizeof(int)*nLenB);
     }
 
     nMaxLen = 0;
-    for(int i = 0; i < nLenA; i++)
+    for(size_t i = 0; i < nLenA; i++)
     {
-        for(int j = 0; j < nLenB; j++)
+        for(size_t j = 0; j < nLenB; j++)
         {
             if(szA[i] == szB[j])
             {
@@ -48,7 +48,7 @@ int Problem05582(void)
     }
     printf("%d\n", nMaxLen);
     
-    for(int i = 0; i < nLenA; i++)
+    for(size_t i = 0; i < nLenA; i++)
     {
         free(pp_nAB_DP[i]);
     }
