+++ ACMICPC/ACMICPC/09000-10000/09252.c
@@ -20,16 +20,16 @@ int Problem09252(void)
     nLenB = strlen(szB);
 
     pp_nAB_DP = (int**)malloc(sizeof(int *)*(nLenA + 1));
-    for(int i = 0; i <= nLenA; i++)
+    for(size_t i = 0; i <= nLenA; i++)
     {
         pp_nAB_DP[i] = (int*)malloc(sizeof(int)*(nLenB + 1));
         memset(pp_nAB_DP[i], 0, sizeof(int)*(nLenB + 1));
     }
 
     nLenResult = 0;
-    for(int i = 1; i < nLenA; i++)
+    for(size_t i = 1; i < nLenA; i++)
     {
-        for(int j = 1; j < nLenB; j++)
+        for(size_t j = 1; j < nLenB; j++)
         {
             if(szA[i] == szB[j])
             {
@@ -62,7 +62,7 @@ int Problem09252(void)
         }
     }
     printf("%s\n",szResult);
-    for(int i = 0; i <= nLenA; i++)
+    for(size_t i = 0; i <= nLenA; i++)
     {
         free(pp_nAB_DP[i]);
     }
