+++ ACMICPC/ACMICPC/10000-11000/10809.c
@@ -16,7 +16,7 @@ int Problem10809(void)
         a_nCount[i] = -1;
     }
 
-    for(int i = 0; i < llnStringLen; i++)
+    for(size_t i = 0; i < llnStringLen; i++)
     {
         if(a_nCount[szString[i] - 'a'] == -1)
             a_nCount[szString[i] - 'a'] = i;
